Move "Get vaccination slots" to common endpoint
