+++ app/Controllers/Patient/VaccinationController.cs
@@ -26,14 +26,6 @@ namespace backend.Controllers.Patient
             return Ok(await this.vaccinationService.ShowAvailableVaccines(request));
         }
 
-        // GET patient/vaccination-slots
-        [HttpGet("vaccination-slots")]
-        [Authorize(AccountTypeEnum.Patient)]
-        public async Task<IActionResult> GetAvailableVaccinationSlots()
-        {
-            return Ok(await this.vaccinationService.GetAvailableVaccinationSlots());
-        }
-
         // PUT patient/vaccination-slots/:{vaccinationSlotId}
         [HttpPut("vaccination-slots/{vaccinationSlotId:int}")]
         [Authorize(AccountTypeEnum.Patient)]
